requirements.txt
Django~=4.2.5
pytz
djangorestframework~=3.14.0
django-rest-passwordreset
psycopg2-binary
pyyaml~=6.0.1
ujson~=5.8.0
requests
black
celery~=5.3.4
drf-spectacular
redis
django-allauth
django-baton
easy-thumbnails
pytest-django